+++ src/rgr.rs
@@ -23,6 +23,7 @@ fn main() -> anyhow::Result<()> {
         .subcommand(cmd_rgr::replace::make_subcommand())
         .subcommand(cmd_rgr::runlist::make_subcommand())
         .subcommand(cmd_rgr::select::make_subcommand())
+        .subcommand(cmd_rgr::span::make_subcommand())
         .subcommand(cmd_rgr::sort::make_subcommand())
         .after_help(
             r###"
@@ -50,6 +51,7 @@ Subcommand groups:
         Some(("replace", sub_matches)) => cmd_rgr::replace::execute(sub_matches),
         Some(("runlist", sub_matches)) => cmd_rgr::runlist::execute(sub_matches),
         Some(("select", sub_matches)) => cmd_rgr::select::execute(sub_matches),
+        Some(("span", sub_matches)) => cmd_rgr::span::execute(sub_matches),
         Some(("sort", sub_matches)) => cmd_rgr::sort::execute(sub_matches),
         _ => unreachable!(),
     }