+++ src/cmd_rgr/mod.rs
@@ -4,6 +4,7 @@ pub mod count;
 pub mod dedup;
 pub mod field;
 pub mod filter;
+pub mod keep;
 pub mod md;
 pub mod merge;
 pub mod pl_2rmp;