+++ src/cmd_rgr/filter.rs
@@ -1,5 +1,4 @@
 use clap::*;
-use std::collections::BTreeMap;
 use std::io::BufRead;
 
 // Create clap subcommand arguments
@@ -69,6 +68,42 @@ Examples:
                 .help("Filter lines by field:STR, FIELD != STR"),
         )
         .arg(
+            Arg::new("eq")
+                .long("eq")
+                .action(ArgAction::Append)
+                .help("Filter lines by field:NUM, FIELD == NUM"),
+        )
+        .arg(
+            Arg::new("ne")
+                .long("ne")
+                .action(ArgAction::Append)
+                .help("Filter lines by field:NUM, FIELD != NUM"),
+        )
+        .arg(
+            Arg::new("lt")
+                .long("lt")
+                .action(ArgAction::Append)
+                .help("Filter lines by field:NUM, FIELD < NUM"),
+        )
+        .arg(
+            Arg::new("le")
+                .long("le")
+                .action(ArgAction::Append)
+                .help("Filter lines by field:NUM, FIELD <= NUM"),
+        )
+        .arg(
+            Arg::new("gt")
+                .long("gt")
+                .action(ArgAction::Append)
+                .help("Filter lines by field:NUM, FIELD > NUM"),
+        )
+        .arg(
+            Arg::new("ge")
+                .long("ge")
+                .action(ArgAction::Append)
+                .help("Filter lines by field:NUM, FIELD >= NUM"),
+        )
+        .arg(
             Arg::new("outfile")
                 .long("outfile")
                 .short('o')
@@ -91,8 +126,15 @@ pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
     let is_invert = args.get_flag("invert");
     let is_insensitive = args.get_flag("case");
 
-    let str_eq_of = opt_idx_str(args, "str-eq", is_insensitive);
-    let str_ne_of = opt_idx_str(args, "str-ne", is_insensitive);
+    let str_eq_tpl = opt_fields_str(args, "str-eq", is_insensitive);
+    let str_ne_tpl = opt_fields_str(args, "str-ne", is_insensitive);
+
+    let num_eq_tpl = opt_fields_num(args, "eq");
+    let num_ne_tpl = opt_fields_num(args, "ne");
+    let num_lt_tpl = opt_fields_num(args, "lt");
+    let num_le_tpl = opt_fields_num(args, "le");
+    let num_gt_tpl = opt_fields_num(args, "gt");
+    let num_ge_tpl = opt_fields_num(args, "ge");
 
     //----------------------------
     // Ops
@@ -119,8 +161,8 @@ pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
             let mut pass_ary = vec![];
 
             // --str-eq and --str-ne
-            if !str_eq_of.is_empty() {
-                for (k, v) in &str_eq_of {
+            if !str_eq_tpl.is_empty() {
+                for (k, v) in &str_eq_tpl {
                     let val = parts.get(k - 1).unwrap();
                     let pass = if is_insensitive {
                         *val.to_ascii_uppercase() == *v
@@ -130,14 +172,58 @@ pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
                     pass_ary.push(pass);
                 }
             }
-            if !str_ne_of.is_empty() {
-                for (k, v) in &str_ne_of {
+            if !str_ne_tpl.is_empty() {
+                for (k, v) in &str_ne_tpl {
                     let val = parts.get(k - 1).unwrap();
                     let pass = if is_insensitive {
                         *val.to_ascii_uppercase() != *v
                     } else {
                         *val != *v
                     };
+                    pass_ary.push(pass);
+                }
+            }
+
+            // --eq, --ne, --gt and --ge
+            if !num_eq_tpl.is_empty() {
+                for (k, v) in &num_eq_tpl {
+                    let val = parts.get(k - 1).unwrap().parse::<f64>().unwrap();
+                    let pass = val == *v;
+                    pass_ary.push(pass);
+                }
+            }
+            if !num_ne_tpl.is_empty() {
+                for (k, v) in &num_ne_tpl {
+                    let val = parts.get(k - 1).unwrap().parse::<f64>().unwrap();
+                    let pass = val != *v;
+                    pass_ary.push(pass);
+                }
+            }
+            if !num_lt_tpl.is_empty() {
+                for (k, v) in &num_lt_tpl {
+                    let val = parts.get(k - 1).unwrap().parse::<f64>().unwrap();
+                    let pass = val < *v;
+                    pass_ary.push(pass);
+                }
+            }
+            if !num_le_tpl.is_empty() {
+                for (k, v) in &num_le_tpl {
+                    let val = parts.get(k - 1).unwrap().parse::<f64>().unwrap();
+                    let pass = val <= *v;
+                    pass_ary.push(pass);
+                }
+            }
+            if !num_gt_tpl.is_empty() {
+                for (k, v) in &num_gt_tpl {
+                    let val = parts.get(k - 1).unwrap().parse::<f64>().unwrap();
+                    let pass = val > *v;
+                    pass_ary.push(pass);
+                }
+            }
+            if !num_ge_tpl.is_empty() {
+                for (k, v) in &num_ge_tpl {
+                    let val = parts.get(k - 1).unwrap().parse::<f64>().unwrap();
+                    let pass = val >= *v;
                     pass_ary.push(pass);
                 }
             }
@@ -164,8 +250,8 @@ pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
     Ok(())
 }
 
-fn opt_idx_str(args: &ArgMatches, id: &str, is_insensitive: bool) -> BTreeMap<usize, String> {
-    let mut str_eq_of: BTreeMap<usize, String> = BTreeMap::new();
+fn opt_fields_str(args: &ArgMatches, id: &str, is_insensitive: bool) -> Vec<(usize, String)> {
+    let mut str_cmp_tpl: Vec<(usize, String)> = Vec::new();
     if args.contains_id(id) {
         for s in args.get_many::<String>(id).unwrap() {
             let parts: Vec<&str> = s.splitn(2, ':').collect();
@@ -175,14 +261,38 @@ fn opt_idx_str(args: &ArgMatches, id: &str, is_insensitive: bool) -> BTreeMap<us
                 std::process::exit(1)
             }
 
-            let idx = parts.first().unwrap().parse::<usize>().unwrap();
+            let fields = intspan::fields_to_idx(parts.first().unwrap());
 
-            if is_insensitive {
-                str_eq_of.insert(idx, parts.get(1).unwrap().to_string().to_ascii_uppercase());
-            } else {
-                str_eq_of.insert(idx, parts.get(1).unwrap().to_string());
+            for idx in &fields {
+                if is_insensitive {
+                    str_cmp_tpl
+                        .push((*idx, parts.get(1).unwrap().to_string().to_ascii_uppercase()));
+                } else {
+                    str_cmp_tpl.push((*idx, parts.get(1).unwrap().to_string()));
+                }
             }
         }
     }
-    str_eq_of
+    str_cmp_tpl
+}
+
+fn opt_fields_num(args: &ArgMatches, id: &str) -> Vec<(usize, f64)> {
+    let mut num_cmp_tpl: Vec<(usize, f64)> = Vec::new();
+    if args.contains_id(id) {
+        for s in args.get_many::<String>(id).unwrap() {
+            let parts: Vec<&str> = s.splitn(2, ':').collect();
+
+            if parts.len() != 2 {
+                eprintln!("Need a valid value for --{} {}", id, s);
+                std::process::exit(1)
+            }
+
+            let fields = intspan::fields_to_idx(parts.first().unwrap());
+
+            for idx in &fields {
+                num_cmp_tpl.push((*idx, parts.get(1).unwrap().parse::<f64>().unwrap()));
+            }
+        }
+    }
+    num_cmp_tpl
 }