+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased - ReleaseDate
 
+## 0.8.3 - 2024-12-29
+
 * Add `rgr span`
 * Add `rgr keep`